App&Scheduler: various performance improvements
We no longer schedule accelerated labs with regular
lectures and vice versa. The scheduler does not generate
schedules in render() now. We also now cache the generated
schedule in Scheduler to avoid redoing all of our work
when, for example, only the filter is changed, or when only
the sort mode is changed.